+++ src/ScottPlot5/ScottPlot5/Interfaces/IScatterSource.cs
@@ -16,6 +16,12 @@ public interface IScatterSource
     /// </summary>
     DataPoint GetNearest(Coordinates location, RenderDetails renderInfo, float maxDistance = 15);
 
+    /// <summary>
+    /// Return the point nearest a specific X location given the X/Y pixel scaling information from a previous render.
+    /// Will return <see cref="DataPoint.None"/> if the nearest point is greater than <paramref name="maxDistance"/> pixels away.
+    /// </summary>
+    DataPoint GetNearestX(Coordinates location, RenderDetails renderInfo, float maxDistance = 15);
+
     public CoordinateRange GetLimitsX();
     public CoordinateRange GetLimitsY();